Improve the model-viewer shader code
These changes refactor the shading code to have a slightly less naive design for the `IMaterial` interface, so that it uses an associated type for the BRDF, etc. While the shading is still completely naive (just Blinn-Phong), the actual decomposition is close to what was done in our Falcor refactoring effort.

I went ahead and spruced up the `cube.obj` file so that it uses more than one material, although it is still obviously a completely uninteresting model to show.